create generic class for setup and teardown
