+++ tests/fixtures.py
@@ -1,0 +1,28 @@
+from cms.api import add_plugin, create_page
+from cms.test_utils.testcases import CMSTestCase
+
+
+class B4TestFixture(object):
+    """Sets up generic setUp and tearDown methods for tests."""
+
+    def setUp(self):
+        self.language = "en"
+        self.home = create_page(
+            title="home",
+            template="page.html",
+            language=self.language,
+        )
+        self.home.publish(self.language)
+        self.page = create_page(
+            title="content",
+            template="page.html",
+            language=self.language,
+        )
+        self.page.publish(self.language)
+        self.placeholder = self.page.placeholders.get(slot="content")
+        self.superuser = self.get_superuser()
+
+    def tearDown(self):
+        self.page.delete()
+        self.home.delete()
+        self.superuser.delete()
